requirements.txt
websocket-client
numpy
newtulipy
python-dotenv
pandas
kaleido
pytest
tk
ttkthemes
Cerberus
